created projects card component
