+++ src/components/index.js
@@ -1,2 +1,3 @@
 export { default as Nav } from './Nav/Nav'
 export { default as SkillsCard } from './SkillsCard/SkillsCard'
+export { default as ProjectsCard } from './ProjectsCard/ProjectsCard'
