+++ src/pages/Projects/Projects.module.css
@@ -65,6 +65,10 @@
     display: inline;
 }
 
+.project1Text {
+    display: inline-block;
+}
+
 /* .projectButton {
     display: none;
     width: 10rem;
